Workbench-Backends/Workbench-ESA-Backend/requirements.txt
Flask~=1.0.2
numpy~=1.15.2
pandas~=0.24.2
nltk~=3.4
requests~=2.20.0
Werkzeug~=0.14.1
mysql-connector-python~=8.0.23
workbench-tools